# frontend/src/lib/api/openapidoc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "backend | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    }
  ],
  "paths": {
...
    "/api/v1/User/{username}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/User/edit/self": {
      "patch": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEditDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEditDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserEditDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/User/edit/{userid}": {
      "patch": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "userid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEditDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEditDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserEditDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "localUsername": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "department": {
            "type": "string",
            "nullable": true
          },
          "trainingOccupation": {
            "type": "string",
            "nullable": true
          },
          "trainingStartYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "UserEditDTO": {
        "type": "object",
        "properties": {
          "localUsername": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "location": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "department": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "trainingOccupation": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "trainingStartYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      }
    }
  },
...
}